refactor: Convexを Feature Slices + CQRS + Policy Patternに移行
- shop/, user/, invite/ ディレクトリに分割
- 各ドメインに queries.ts, mutations.ts, policies.ts を配置
- 旧ファイル (shop.ts, user.ts, invite.ts) を削除
- ロール判定ロジックを policies.ts に集約
